requirements.txt
fastapi
uvicorn[standard]
python-dotenv
volcengine
aiohttp
